canvas: ParticleTextCanvas + particle-text action
WebGL2 transform-feedback particle system ported to Svelte action.
The action takes container+textCanvas refs as params (multi-canvas
coordination: text bitmap drawn on a 2D canvas, particles on a WebGL2
canvas on top). Action handles context-loss/restore for iOS Safari.

ParticleTextCanvas uses bind:this for the two internal canvas refs
(autofixer noted this; kept intentionally — action needs both nodes).
$derived builds glParams only after both refs are non-null.
